% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logsumexp.R
\name{helper_functions}
\alias{helper_functions}
\alias{logsumexp}
\title{A stable function for computing log(sum(exp(x)))}
\usage{
logsumexp(x)
}
\arguments{
\item{x}{The values for which you want to compute log(sum(exp(x)))}
}
\value{
The stable result of log(sum(exp(x)))
}
\description{
A stable function for computing log(sum(exp(x)))
}
