% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logsumexp.R
\name{logsumexp}
\alias{logsumexp}
\title{Stable log sum of exponential calculations}
\usage{
logsumexp(x)
}
\arguments{
\item{x}{The values for which you want to compute log(sum(exp(x)))}
}
\value{
The stable result of log(sum(exp(x)))
}
\description{
The function \code{logsumexp} is used for stable computation of log(sum(exp(x))), which is useful when summing weights for example.
}
\seealso{
See \link{ZVCV} for more package details.
}
